refactor(search): split search-registry.ts into 4 modules behind a barrel
src/core/search-registry.ts (1,154 lines) → a 16-line barrel re-exporting four focused
modules:
- search-registry-core.ts (444) — globalThis registry state + lifecycle + table
  provisioning + resolvers + shared query/result mappers. Owns the HMR state + reset.
- search-plugin-api.ts (413) — buildSearchAPI + the startup reconcile drain.
- search-reindex.ts (268) — reindexContentTypes + getSearchHealth.
- search-query.ts (102) — crossTableSearch.

Barrel strategy keeps every consumer import, all 11 test files (including the 7 that
mock.module('@/core/search-registry')), and server.ts's dynamic-import strings UNCHANGED —
only internal cross-module wiring uses the sub-module paths. No cycles: barrel → all four;
feature modules → core (plugin-api also → reindex for api.health); core imports none of
them. Logger pretty-info allowlist extended with the four new module names. Pure relocation:
the pluginTables primary-table fix is preserved verbatim.

Verified: typecheck + lint clean; 149/0 across the 11 search test files; full suite 5105/0;
full 3-target binary build; and an isolated-home boot smoke of the built binary against a
real Antfly — tables register, plugins activate through the split modules, and team's
/search resolves its primary table with no throw (the path that previously threw for a
multi-content plugin).
